+++ CondemnedAssistance/Controllers/HelpController.cs
@@ -1,5 +1,7 @@
 ﻿using CondemnedAssistance.Helpers;
 using CondemnedAssistance.Models;
+using CondemnedAssistance.Services.Security._Constants;
+using CondemnedAssistance.Services.Security.Help;
 using CondemnedAssistance.ViewModels;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
@@ -9,15 +11,18 @@ using System.Linq;
 using System.Threading.Tasks;
 
 namespace CondemnedAssistance.Controllers {
-    [Authorize(Roles = "3")]
     public class HelpController : Microsoft.AspNetCore.Mvc.Controller {
 
         private UserContext _db;
         private LinkHelper linkHelper;
+        private IAuthorizationService _authorizationService;
+        private int _controllerId;
 
-        public HelpController(UserContext context) {
+        public HelpController(UserContext context, IAuthorizationService authorizationService) {
             _db = context;
             linkHelper = new LinkHelper(context, "userEdit");
+            _authorizationService = authorizationService;
+            _controllerId = _db.Controllers.Single(c => c.NormalizedName == Constants.Help).Id;
         }
 
         [HttpGet]
@@ -67,19 +72,35 @@ namespace CondemnedAssistance.Controllers {
         {
             return View();
         }
+
         [HttpGet]
-        public IActionResult Index() {
+        public async Task<IActionResult> Index() {
+            AuthorizationResult result = await _authorizationService.AuthorizeAsync(User, _controllerId, HelpOperations.Read);
+
+            if (!result.Succeeded) {
+                return new ChallengeResult();
+            }
             List<Help> model = _db.Helps.ToList();
             return View(model);
         }
 
         [HttpGet]
-        public IActionResult Create() {
+        public async Task<IActionResult> Create() {
+            AuthorizationResult result = await _authorizationService.AuthorizeAsync(User, _controllerId, HelpOperations.Create);
+
+            if (!result.Succeeded) {
+                return new ChallengeResult();
+            }
             return View();
         }
 
         [HttpPost]
-        public IActionResult Create(Help model) {
+        public async Task<IActionResult> Create(Help model) {
+            AuthorizationResult result = await _authorizationService.AuthorizeAsync(User, _controllerId, HelpOperations.Create);
+
+            if (!result.Succeeded) {
+                return new ChallengeResult();
+            }
             if (ModelState.IsValid) {
                 if(!_db.Helps.Any(h => h.NormalizedName == model.Name.ToUpper())) {
                     model.NormalizedName = model.Name.ToUpper();
@@ -96,13 +117,23 @@ namespace CondemnedAssistance.Controllers {
         }
 
         [HttpGet]
-        public IActionResult Update(int id) {
+        public async Task<IActionResult> Update(int id) {
+            AuthorizationResult result = await _authorizationService.AuthorizeAsync(User, _controllerId, HelpOperations.Update);
+
+            if (!result.Succeeded) {
+                return new ChallengeResult();
+            }
             Help model = _db.Helps.Single(h => h.Id == id);
             return View(model);
         }
 
         [HttpPost]
-        public IActionResult Update(int id, Help model) {
+        public async Task<IActionResult> Update(int id, Help model) {
+            AuthorizationResult result = await _authorizationService.AuthorizeAsync(User, _controllerId, HelpOperations.Update);
+
+            if (!result.Succeeded) {
+                return new ChallengeResult();
+            }
             if (ModelState.IsValid) {
                 if(!_db.Helps.Any(h => h.NormalizedName == model.Name.ToUpper() && h.Id != id)) {
                     model.NormalizedName = model.Name.ToUpper();
@@ -120,7 +151,12 @@ namespace CondemnedAssistance.Controllers {
         }
 
         [HttpGet]
-        public IActionResult Delete(int id) {
+        public async Task<IActionResult> Delete(int id) {
+            AuthorizationResult result = await _authorizationService.AuthorizeAsync(User, _controllerId, HelpOperations.Delete);
+
+            if (!result.Succeeded) {
+                return new ChallengeResult();
+            }
             if (_db.Helps.Any(h => h.Id == id)) {
                 Help help = _db.Helps.Single(h => h.Id == id);
                 _db.Helps.Remove(help);
@@ -130,14 +166,23 @@ namespace CondemnedAssistance.Controllers {
         }
 
         [HttpGet]
-        public IActionResult UserHelpList(int userId) {
+        public async Task<IActionResult> UserHelpList(int userId) {
+            AuthorizationResult result = await _authorizationService.AuthorizeAsync(User, _controllerId, HelpOperations.UserHelpList);
+
+            if (!result.Succeeded) {
+                return new ChallengeResult();
+            }
             int[] userHelpList = _db.UserHelps.Where(h => h.UserId == userId).Select(h => h.HelpId).ToArray();
             return View(_db.Helps.Where(h => userHelpList.Contains(h.Id)).ToList());
         }
 
         [HttpGet]
-        public IActionResult AddUserHelp(int userId) {
+        public async Task<IActionResult> AddUserHelp(int userId) {
+            AuthorizationResult result = await _authorizationService.AuthorizeAsync(User, _controllerId, HelpOperations.AddUserHelp);
 
+            if (!result.Succeeded) {
+                return new ChallengeResult();
+            }
             Dictionary<string, string> routeVals = new Dictionary<string, string> { };
             routeVals.Add("id", userId.ToString());
 
@@ -177,7 +222,12 @@ namespace CondemnedAssistance.Controllers {
         }
 
         [HttpPost]
-        public IActionResult AddUserHelp(int userId, HelpModel model) {
+        public async Task<IActionResult> AddUserHelp(int userId, HelpModel model) {
+            AuthorizationResult result = await _authorizationService.AuthorizeAsync(User, _controllerId, HelpOperations.AddUserHelp);
+
+            if (!result.Succeeded) {
+                return new ChallengeResult();
+            }
             if (ModelState.IsValid) {
                 if (_db.UserHelps.Any(h => h.UserId == userId)) {
                     foreach (UserHelp help in _db.UserHelps.Where(h => h.UserId == userId)) {
